% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.r
\name{augmentation_norm}
\alias{augmentation_norm}
\title{Data Augmentation with Normal Distribution}
\usage{
augmentation_norm(ocr)
}
\arguments{
\item{ocr}{A GRanges object with a 'thick' column of IRanges.}
}
\value{
A GRanges object with normally shifted ranges.
}
\description{
This function applies normal distribution shifts to genomic ranges in the
given GRanges object and filters the results to avoid overlaps.
}
\examples{
# ocr is a GRanges object with required data
augmented_ocr <- augmentation_norm(ocr)
}
