% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.r
\name{augmentation_unif}
\alias{augmentation_unif}
\title{Data Augmentation with Uniform Distribution}
\usage{
augmentation_unif(ocr)
}
\arguments{
\item{ocr}{A GRanges object with a 'thick' column of IRanges.}
}
\value{
A GRanges object with uniformly shifted ranges.
}
\description{
This function applies uniform distribution shifts to genomic ranges in the
given GRanges object and filters the results to avoid overlaps.
}
\examples{
# ocr is a GRanges object with required data
augmented_ocr <- augmentation_unif(ocr)
}
