% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{cast_rse_to_granges}
\alias{cast_rse_to_granges}
\title{Convert SummarizedExperiment to GRanges with Assay Data}
\usage{
cast_rse_to_granges(rse, assay = "counts", coln_assay = 1, colname = "score")
}
\arguments{
\item{rse}{A `SummarizedExperiment` object containing the genomic ranges and
assay data.}

\item{assay}{A character string specifying the assay name to extract from
the `SummarizedExperiment` object. Default is "counts".}

\item{coln_assay}{An integer specifying the column index of the assay to use.
Default is 1.}

\item{colname}{A character string specifying the name of the column to be
added to the `GRanges` object. Default is "score".}
}
\value{
A `GRanges` object with the specified assay data added as a metadata
  column.
}
\description{
This function converts a `SummarizedExperiment` object to a `GRanges` object
and adds a specified assay data column to the resulting `GRanges` object.
}
\examples{
# Example usage:
# rse is a SummarizedExperiment object
gr <- cast_rse_to_granges(rse, assay = "counts",
                          coln_assay = 1, colname = "score")
}
