% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.r
\name{combine_plus_minus_profiles}
\alias{combine_plus_minus_profiles}
\title{Combine Plus and Minus Strand Profiles}
\usage{
combine_plus_minus_profiles(count_profiles, len_vec)
}
\arguments{
\item{count_profiles}{A list containing count profiles for both forward (`+`)
and reverse (`-`) strands.}

\item{len_vec}{An integer specifying the length of the profiles.}
}
\value{
A data frame containing combined profiles with modified row names.
}
\description{
This function combines forward (plus) and reverse (minus) strand count
profiles and modifies their row names.
}
