% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_helper.r
\name{convert_rowname_to_nostrand}
\alias{convert_rowname_to_nostrand}
\title{Convert Row Name Strands to No-strand}
\usage{
convert_rowname_to_nostrand(strand_str)
}
\arguments{
\item{strand_str}{A character vector of row names
containing strand information.}
}
\value{
A character vector with strand information converted to "*".
}
\description{
This function converts the strand information
in row names from "+" or "-" to "*".
}
