% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_gr.r
\name{convert_strand_to_nostrand_gr}
\alias{convert_strand_to_nostrand_gr}
\title{Convert Strand Information to No Strand for GRanges Object}
\usage{
convert_strand_to_nostrand_gr(region_gr)
}
\arguments{
\item{region_gr}{A `GRanges` object containing genomic ranges.}
}
\value{
A `GRanges` object with the strand information set to `"*"`.
}
\description{
This function takes a `GRanges` object
and sets the strand information to `"*"`,
indicating no strand specificity.
}
\details{
The function modifies the strand information
of the input `GRanges` object, setting it to `"*"`.
This is useful in situations where strand information
is irrelevant or not needed.
}
