% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_helper.r
\name{create_granges_from_rownames}
\alias{create_granges_from_rownames}
\title{Create GRanges Object from Row Names}
\usage{
create_granges_from_rownames(row_names_str)
}
\arguments{
\item{row_names_str}{A character vector of row names with each name formatted
as "chr:start-end;strand".}
}
\value{
A `GRanges` object created from the extracted components of
the rownames.
}
\description{
This function creates a `GRanges` object from row names by extracting
the chromosome, start, end, and strand components.
}
