% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_modification.r
\name{drop_chrM_from_grl}
\alias{drop_chrM_from_grl}
\title{Remove 'chrM' from GRangesList}
\usage{
drop_chrM_from_grl(grl)
}
\arguments{
\item{grl}{A `GRangesList` object containing genomic ranges.}
}
\value{
A `GRangesList` object with all ranges on 'chrM' removed.
}
\description{
This function removes all genomic ranges located on the chromosome 'chrM'
from each `GRanges` object within a `GRangesList`. It is useful for excluding
mitochondrial DNA, which is often represented by 'chrM', from analyses.
}
\examples{
library(GenomicRanges)
gr1 <- GRanges(seqnames = c("chr1", "chrM", "chr2"), 
               ranges = IRanges(start = 1:3, width = 3), strand = "+")
gr2 <- GRanges(seqnames = c("chrM", "chr2", "chr3"), 
               ranges = IRanges(start = 4:6, width = 3), strand = "-")
grl <- GRangesList(gr1 = gr1, gr2 = gr2)

# Remove 'chrM' ranges
modified_grl <- drop_chrM_from_grl(grl)
}
