% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_df.r
\name{extend_from_center_thick_df}
\alias{extend_from_center_thick_df}
\title{Extend Intervals from the Center}
\usage{
extend_from_center_thick_df(
  df,
  dis,
  keep_same_length = TRUE,
  group_exact_positions = TRUE
)
}
\arguments{
\item{df}{A data frame containing "start" and "end" columns representing
genomic intervals.}

\item{dis}{A numeric value specifying the distance to extend the intervals
from the center.}

\item{keep_same_length}{A logical value indicating whether to maintain the
same length for all intervals after extension. Defaults to TRUE.}

\item{group_exact_positions}{A logical value indicating whether to group
intervals that have the exact same positions. Defaults to TRUE.}
}
\value{
A data frame with updated "start", "end", and "length" columns, and
  possibly an updated "thick" column.
}
\description{
Extends the genomic intervals from their center points, known as "thick"
positions, by a specified distance. It can optionally maintain the original
interval lengths and group intervals with exactly the same positions.
}
\examples{
df <- data.frame(start = c(10, 20, 30), end = c(15, 25, 35))
extended_df <- extend_from_center_thick_df(df, dis = 5)
extended_df_same_length <- extend_from_center_thick_df(df, dis = 5, keep_same_length = TRUE) # nolint: line_length_linter.
}
