% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_modification.r
\name{extend_from_center_thick_gr}
\alias{extend_from_center_thick_gr}
\title{Extend Genomic Ranges from the Center Based on Thickness}
\usage{
extend_from_center_thick_gr(gr, dis, keep_same_length = TRUE)
}
\arguments{
\item{gr}{A `GRanges` object containing genomic ranges with a "thick" column
in the metadata.}

\item{dis}{A numeric value specifying the distance to extend from the center.}

\item{keep_same_length}{A logical value indicating whether to keep the
intervals at the same length after extension. Defaults to TRUE.}
}
\value{
A `GRanges` object with updated start and end positions, and an
  updated "thick" column if `keep_same_length` is TRUE.
}
\description{
This function extends genomic ranges from their center points, defined by
the "thick" positions in the metadata columns of a `GRanges` object. The
"thick" column must exist and be of class `IRanges`. The extension can
either maintain the original interval lengths or allow them to vary.
}
\examples{
# Assuming `gr` is a GRanges object with a 'thick' column of class IRanges
extended_gr <- extend_from_center_thick_gr(gr, dis = 100, keep_same_length = TRUE) # nolint: line_length_linter.

}
