% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{extract_core}
\alias{extract_core}
\title{Extract core region from GRanges object}
\usage{
extract_core(gr, ext_core = 75)
}
\arguments{
\item{gr}{A GRanges object representing genomic ranges.}

\item{ext_core}{The number of base pairs to extract
on either side of the center
(default: 75, resulting in a 151 bp total length).}
}
\value{
A GRanges object with the core regions extracted.
}
\description{
This function extracts a core region of specified length
from the center of each range in a GRanges object.
}
