% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_helper.r
\name{extract_rowname_components}
\alias{extract_rowname_components}
\title{Extract Components from Row Names}
\usage{
extract_rowname_components(row_names_cpn)
}
\arguments{
\item{row_names_cpn}{A character vector of row names with each name formatted
as "chr:start-end;strand".}
}
\value{
A character vector containing the chromosome, start, end, and strand
extracted from the row name.
}
\description{
This function extracts the chromosome, start, end, and strand components 
from row names formatted as "chr:start-end;strand".
}
