% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs.r
\name{filter_blacklist}
\alias{filter_blacklist}
\title{Filter Ranges Overlapping Non-reliable Genomic Regions}
\usage{
filter_blacklist(ranges, blacklist)
}
\arguments{
\item{ranges}{A GRanges object containing the ranges to be filtered.}

\item{blacklist}{A GRanges object containing the blacklist of non-reliable
genomic regions.}
}
\value{
A GRanges object with ranges that do not overlap the blacklist.
}
\description{
Filters out ranges from a GRanges object that overlap with a specified
blacklist of non-reliable genomic regions.
}
\examples{
# Example usage:
# ranges <- GRanges(seqnames=Rle(c("chr1", "chr2")),
#                   ranges=IRanges(start=c(1, 3), end=c(5, 7)))
# blacklist <- GRanges(seqnames=Rle(c("chr1")),
#                      ranges=IRanges(start=c(2), end=c(4)))
# filtered_ranges <- filter_blacklist(ranges, blacklist)
}
