% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonoverlapping_ranges.r
\name{get_highest_non_overlap}
\alias{get_highest_non_overlap}
\title{Get Highest Non-Overlapping Ranges for All Chromosomes in Parallel}
\usage{
get_highest_non_overlap(gr, num_cores = parallel::detectCores() - 1)
}
\arguments{
\item{gr}{A GenomicRanges::GRanges object with a 'score' metadata column.}

\item{num_cores}{The number of cores to use for parallel processing.
Default is parallel::detectCores() - 1.}
}
\value{
A GenomicRanges::GRanges object with
the highest non-overlapping ranges for all chromosomes.
}
\description{
This function takes a GenomicRanges::GRanges object
and returns the highest non-overlapping ranges
for all chromosomes, processed in parallel.
}
