% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonoverlapping_ranges.r
\name{get_highest_non_overlap_chr}
\alias{get_highest_non_overlap_chr}
\title{Get Highest Non-Overlapping Ranges for a Chromosome}
\usage{
get_highest_non_overlap_chr(gr)
}
\arguments{
\item{gr}{A GenomicRanges::GRanges object with a 'score' metadata column.}
}
\value{
A GenomicRanges::GRanges object with
the highest non-overlapping ranges.
}
\description{
This function takes a GenomicRanges::GRanges object
and returns the highest non-overlapping ranges
based on the scores in the metadata column.
}
