% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata for each GRanges object.}
\usage{
get_metadata(gr_range, filtered_gr)
}
\arguments{
\item{gr_range}{A single GRanges range.}

\item{filtered_gr}{A GRanges object with associated scores.}
}
\value{
A list containing metadata: thick position,
maximum score, and all scores.
}
\description{
This function calculates metadata for each genomic range,
including the thick position, maximum score, and all scores in the range.
}
