% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.r
\name{get_random_number}
\alias{get_random_number}
\title{Generate Random Numbers with Uniform and Normal Distributions}
\usage{
get_random_number(n)
}
\arguments{
\item{n}{An integer specifying the number of random numbers to generate.}
}
\value{
A list containing two vectors: "unif" for uniform distribution and
"norm" for normal distribution.
}
\description{
This function generates random integers from both uniform and normal
distributions and plots their histograms.
}
\examples{
get_random_number(10)
}
