% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_modification.r
\name{group_exact_positions_df}
\alias{group_exact_positions_df}
\title{Group Genomic Intervals by Exact Positions}
\usage{
group_exact_positions_df(df)
}
\arguments{
\item{df}{A data frame containing genomic intervals with "seqnames",
"start", and "end" columns.}
}
\value{
A data frame with intervals grouped by identical positions, with
  concatenated unique values from other columns.
}
\description{
This function groups genomic intervals in a DataFrame that have identical
"seqnames", "start", and "end" positions. Optionally, the "thick" column can
also be considered if present. For each group of intervals with the same
positions, the function concatenates unique values from other columns, using
a semicolon as the separator.
}
\examples{
df <- data.frame(seqnames = c("chr1", "chr1", "chr2", "chr2"),
                 start = c(100, 100, 200, 200),
                 end = c(150, 150, 250, 250),
                 value = c("A", "B", "C", "D"))
grouped_df <- group_exact_positions_df(df)
}
