% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_split_data_basedon_annotation.r
\name{keep_equal_num}
\alias{keep_equal_num}
\title{Keep Equal Number of Ranges}
\usage{
keep_equal_num(gr, threshold)
}
\arguments{
\item{gr}{A `GRanges` object containing genomic ranges.}

\item{threshold}{An integer specifying the number of ranges to select.}
}
\value{
A `GRanges` object containing the randomly selected ranges, sorted
  by genomic coordinates.
}
\description{
This function randomly selects a specified number of genomic ranges from a
`GRanges` object to ensure equal representation.
}
\examples{
gr <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges::IRanges(1:10, width = 1))
reduced_gr <- keep_equal_num(gr, 5)
}
