% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.r
\name{modify_profile_rownames}
\alias{modify_profile_rownames}
\title{Modify Profile Row Names}
\usage{
modify_profile_rownames(profiles, count_profiles)
}
\arguments{
\item{profiles}{A data frame of profile data where row names represent
genomic coordinates.}

\item{count_profiles}{A list containing count profiles for both forward (`+`)
and reverse (`-`) strands.}
}
\value{
A data frame with modified row names
if the forward and reverse strand lists
are identical after strand conversion.
If the lists are not identical,
a message is printed, and the row names are not modified.
}
\description{
This function modifies the row names of profile data by converting strand
information to no-strand and ensuring consistency between forward and
reverse strands.
}
