% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.r
\name{PRIMEloci_profile}
\alias{PRIMEloci_profile}
\title{Main function to process profiles for each chromosome and save results}
\usage{
PRIMEloci_profile(
  ctss_rse,
  regions_gr,
  output_dir,
  output_dir_name,
  ext_dis,
  addtn_to_filename = "",
  save_count_profiles = FALSE,
  file_type = "parquet",
  num_cores = NULL
)
}
\arguments{
\item{ctss_rse}{A RangedSummarizedExperiment object containing CTSS counts.}

\item{regions_gr}{A GRanges or GRangesList object containing genomic regions.}

\item{output_dir}{The directory where output files will be saved.}

\item{output_dir_name}{The directory name for the output.}

\item{ext_dis}{Integer value for extending the distance
used in profile computations.}

\item{addtn_to_filename}{A string to add to the output filename.}

\item{save_count_profiles}{Logical flag indicating whether count profiles
should be saved.}

\item{file_type}{The format in which to save the files ("csv" or "parquet").}

\item{num_cores}{The number of cores to use for parallel processing.
If not provided, it defaults to using all available cores minus one.}
}
\description{
This function processes the profiles for all chromosomes using a GRanges
object, computes the count profiles, and calls a chromosome-specific function
to handle individual chromosomes and save the results. The filenames are
generated and passed to the chromosome-specific function.
}
\details{
If `regions_gr` is a `GRangesList`, the function expects paired profiles
between `ctss_rse` and `regions_gr` such that for each element `i`, the
`ctss_rse` column and the corresponding element of `regions_gr` will be
processed together. It is assumed that the order of the elements in
`regions_gr` corresponds to the order in `ctss_rse`.
}
