% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_2.r
\name{PRIMEloci_profile_2}
\alias{PRIMEloci_profile_2}
\title{Process profiles for each column in the CTSS dataset and save results}
\usage{
PRIMEloci_profile_2(
  ctss_rse,
  regions_gr,
  output_dir,
  output_dir_name,
  ext_dis,
  addtn_to_filename = "",
  save_count_profiles = FALSE,
  file_type = "parquet",
  num_cores = NULL
)
}
\arguments{
\item{ctss_rse}{A RangedSummarizedExperiment object containing CTSS counts.}

\item{regions_gr}{A GRanges or GRangesList object containing genomic regions.}

\item{output_dir}{The directory where output files will be saved.}

\item{output_dir_name}{The directory name for the output.}

\item{ext_dis}{Integer value for extending the distance
used in profile computations.}

\item{addtn_to_filename}{A string to add to the output filename.}

\item{save_count_profiles}{Logical flag indicating
whether count profiles should be saved.}

\item{file_type}{The format in which to save the files ("csv" or "parquet").}

\item{num_cores}{The number of cores to use for parallel processing.
If not provided, it defaults to using all available cores minus one.}
}
\description{
This function processes the profiles for each column
in a RangedSummarizedExperiment object (`ctss_rse`),
computes the count profiles, and calls a chromosome-specific function
to handle individual chromosomes and save the results.
}
