% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_PRIMEloci.r
\name{process_all_files}
\alias{process_all_files}
\title{Process All Prediction Files}
\usage{
process_all_files(prediction_dir, partial_name)
}
\arguments{
\item{prediction_dir}{Character.
The directory where the prediction files are located.}

\item{partial_name}{Character.
A pattern to match files in the `prediction_dir`.}
}
\value{
A `GRangesList` object containing the processed results
from all matching files.
}
\description{
This function processes all files matching
a specified pattern in a given directory.
It reads each file, applies the `wrapup_filter_bed_to_reduce` function
from the `PRIMEloci` package,
and stores the results in a `GRangesList`.
}
\details{
The function lists all files in the specified directory
that match the given pattern.
For each file, it applies the `wrapup_filter_bed_to_reduce` function
and stores the result in a `GRangesList` object.
If no files are found matching the pattern, the function stops
and throws an error.
}
\examples{
\dontrun{
gr_list <- process_all_files(
  prediction_dir = "path/to/predictions",
  partial_name = "pattern_to_match"
)
}

}
