% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_2.r
\name{profile_log_message}
\alias{profile_log_message}
\title{Log messages to a file}
\usage{
profile_log_message(message, log_file)
}
\arguments{
\item{message}{A character string containing the log message to write.}

\item{log_file}{A character string specifying
the file path where the log message should be written.}
}
\description{
This function writes log messages to a specified log file,
appending a timestamp to each entry.
}
