% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{report_time_execution}
\alias{report_time_execution}
\title{Measure and Report Execution Time of a Function}
\usage{
report_time_execution(fun)
}
\arguments{
\item{fun}{A function call to be executed and timed. It should be provided 
as an expression or a function object.}
}
\value{
The output of the executed function.
}
\description{
The `report_time_execution` function measures and reports the execution time 
of a given function.
}
\examples{
# Example usage:
# Measure the time taken to run a simple calculation
report_time_execution({
  Sys.sleep(2)
  2 + 2
})
}
