% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.r
\name{rnorm_int}
\alias{rnorm_int}
\title{Generate Random Integers with Normal Distribution}
\usage{
rnorm_int(number)
}
\arguments{
\item{number}{An integer specifying the number of random integers to generate}
}
\value{
A vector of random integers.
}
\description{
This function generates a specified number of random integers from a normal
distribution, truncated to lie within the range -30 to 30.
}
\examples{
rnorm_int(5)
}
