% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_PRIMEloci.r
\name{run_prediction_python_script}
\alias{run_prediction_python_script}
\title{Run a Python Script for Profile Prediction}
\usage{
run_prediction_python_script(
  script_path,
  script_dir,
  profile_main_dir,
  profile_sub_dir,
  model_path,
  name_prefix,
  threshold = 0.5,
  file_format = "parquet"
)
}
\arguments{
\item{script_path}{Character. The path to the Python script to be executed.}

\item{script_dir}{Character. The directory where the Python script resides.}

\item{profile_main_dir}{Character.
The main directory containing profile data.}

\item{profile_sub_dir}{Character.
The subdirectory within the main profile directory.}

\item{model_path}{Character.
The path to the trained model file to be used for prediction.}

\item{name_prefix}{Character.
A prefix to be added to the output files generated by the Python script.}

\item{threshold}{Numeric.
The probability threshold for classifying predictions. Default is 0.5.}

\item{file_format}{Character.
The format of the profile files (e.g., "parquet", "csv").
Default is "parquet".}
}
\value{
None. This function is used for its side effects of
running the Python script.
}
\description{
This function runs a specified Python script using the `reticulate` package.
}
\details{
The function first checks if the specified Python script exists.
If the script exists, it constructs a command string to pass
the required arguments and runs the script using `reticulate::py_run_string`.
The command includes various parameters like the working directory,
profile directories, model path,
output name prefix, classification threshold, and file format.

If the script does not exist, the function stops and throws an error message.
}
\examples{
\dontrun{
run_prediction_python_script(
  script_path = "path/to/predict_profile_probabilities.py",
  script_dir = "path/to/scripts",
  profile_main_dir = "path/to/profiles/main",
  profile_sub_dir = "subdir_name",
  model_path = "path/to/model.h5",
  name_prefix = "output_prefix",
  threshold = 0.7,
  file_format = "parquet"
)
}

}
