% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_PRIMEloci.r
\name{run_PRIMEloci}
\alias{run_PRIMEloci}
\title{Run the PRIMEloci Workflow}
\usage{
run_PRIMEloci(ctss_rse, tc_grl, config_file = "config_R_PRIMEloci.yaml")
}
\arguments{
\item{ctss_rse}{A `SummarizedExperiment` object containing CTSS data.}

\item{tc_grl}{A `GRangesList` object containing TC data.}

\item{config_file}{Character. Path to the configuration file in YAML format.
Default is "config_R_PRIMEloci.yaml".}
}
\value{
A `GRangesList` containing the processed results.
}
\description{
This function encapsulates the entire PRIMEloci workflow,
including creating profiles, running a Python script for profile prediction,
and processing the results into a `GRangesList`.
}
\details{
This function orchestrates the PRIMEloci workflow, including:
- Setting up necessary directories based on the configuration file.
- Creating profiles using the CTSS and TC data.
- Running a Python script for predicting profile probabilities.
- Processing the output files and returning a `GRangesList` object.

The workflow relies on configuration parameters specified in a YAML file.
}
\examples{
\dontrun{
ctss_rse <- loadRDS("path_to_ctss_rse.rds")
tc_grl <- loadRDS("path_to_tc_grl.rds")
gr_list <- run_PRIMEloci(
  ctss_rse = ctss_rse,
  tc_grl = tc_grl,
  config_file = "config_PRIMEloci.yaml"
)
}
}
