% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_related.r
\name{save_granges_to_bed}
\alias{save_granges_to_bed}
\title{Save a GRanges object to RDS and BED formats}
\usage{
save_granges_to_bed(gr, output_dir, input_basename, bed_file)
}
\arguments{
\item{gr}{A `GRanges` object to be saved.}

\item{output_dir}{Character. The directory where the output files will be saved.}

\item{input_basename}{Character. The base name for the output files.}
}
\value{
None. The function saves the output files to the specified directory.
}
\description{
This function saves a `GRanges` object as an RDS file and converts it to a `data.frame`
for saving as a BED file. The BED file is saved with adjusted columns to match the original format.
}
