% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_core_fn.r
\name{save_granges_to_bed_2}
\alias{save_granges_to_bed_2}
\title{Save the GRanges object as both an RDS file and a BED file.}
\usage{
save_granges_to_bed_2(gr, output_dir, input_basename)
}
\arguments{
\item{gr}{A GRanges object.}

\item{output_dir}{Directory where the output files will be saved.}

\item{input_basename}{The base name for the output files.}
}
\description{
This function saves the GRanges object to an RDS file
and converts it to a BED file format, saving both in the specified directory.
}
