% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_helper.r
\name{save_to_file}
\alias{save_to_file}
\title{Save data to file with an appropriate file extension based on file type}
\usage{
save_to_file(data, suffix, file_type, chr_name, file_path)
}
\arguments{
\item{data}{A data frame or object that can be converted to
a data frame to be saved.}

\item{suffix}{A string to be appended to the file path
(e.g., "_metadata", "_profiles").}

\item{file_type}{The format in which to save the file.
Should be either "csv" or "parquet".}

\item{chr_name}{A string representing the chromosome name
that will be appended to the file name.}

\item{file_path}{The base file path where the file will be saved,
excluding the chromosome name and file extension.}
}
\description{
This function saves the given data either
in CSV or Parquet format based on the specified `file_type`.
The function dynamically determines the correct file extension
(`.csv` or `.parquet`) and appends the chromosome name and suffix
to the provided file path to create the full file name.
}
