% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_split_data_basedon_annotation.r
\name{select_negative_to_match_positive}
\alias{select_negative_to_match_positive}
\title{Select Negative Samples to Match Positive Samples}
\usage{
select_negative_to_match_positive(neg_sltocr_grl, match_pos_grl)
}
\arguments{
\item{neg_sltocr_grl}{A `GRangesList` object containing negative selected
open chromatin regions (OCRs).}

\item{match_pos_grl}{A `GRangesList` object containing positive selected
OCRs to match in terms of sample size.}
}
\value{
A `GRangesList` object containing a subset of the negative samples
  matching the size and distribution of the positive samples.
}
\description{
This function selects a subset of negative samples to match the number of
positive samples in a given `GRangesList`. The selection is stratified by
chromosome (`seqnames`) and transcript type (`txType`) to ensure a balanced
representation.
}
\examples{
# Example usage:
# neg_sltocr_grl and match_pos_grl are GRangesList objects
selected_neg_grl <- select_negative_to_match_positive(neg_sltocr_grl, match_pos_grl)
}
