% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_filter.r
\name{select_ocr_on_condition}
\alias{select_ocr_on_condition}
\title{Select OCRs Based on Specific Conditions}
\usage{
select_ocr_on_condition(cage_rse, ocr_gr)
}
\arguments{
\item{cage_rse}{A `SummarizedExperiment` object containing the CAGE dataset.}

\item{ocr_gr}{A `GRanges` object representing open chromatin regions (OCRs).}
}
\value{
A `GRangesList` object containing the selected OCRs for each library
  based on the specified conditions.
}
\description{
This function selects OCRs (open chromatin regions) from a CAGE dataset
across all libraries based on specified conditions. It uses the
`select_ocr_on_condition_each_lib` function for each library to ensure
consistency and accuracy in the selection process.
}
\details{
This function iterates over all libraries in the CAGE dataset and applies
the selection criteria using `select_ocr_on_condition_each_lib`.
}
\examples{
# Example usage:
cage_rse <- ... # Define the SummarizedExperiment object
ocr_gr <- ... # Define the GRanges object for OCRs
selected_ocr_grl <- select_ocr_on_condition(cage_rse, ocr_gr)
}
