% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_filter.r
\name{select_ocr_on_condition_each_lib}
\alias{select_ocr_on_condition_each_lib}
\title{Select OCRs Based on Specific Conditions for Each Library}
\usage{
select_ocr_on_condition_each_lib(
  cage_rse,
  ocr_gr,
  coln,
  min_score_for_onebase_ocr = 4
)
}
\arguments{
\item{cage_rse}{A `SummarizedExperiment` object containing the CAGE dataset.}

\item{ocr_gr}{A `GRanges` object representing open chromatin regions (OCRs).}

\item{coln}{An integer indicating the column index of the assay to use.}

\item{min_score_for_onebase_ocr}{A numeric value specifying the minimum
score required for one-base overlaps. Defaults to 4.}
}
\value{
A `GRanges` object containing the selected OCRs based on the
  specified conditions.
}
\description{
This function selects OCRs (open chromatin regions) from a CAGE dataset
based on specified conditions for each library. It filters the regions
based on the number of overlaps and scores, ensuring that selected OCRs
meet a minimum score threshold for one-base overlaps.
}
\details{
This function calls `cast_rse_to_granges` from the `general.r` script.
}
\examples{
# Example usage:
cage_rse <- ... # Define the SummarizedExperiment object
ocr_gr <- ... # Define the GRanges object for OCRs
selected_ocr <- select_ocr_on_condition_each_lib(cage_rse, ocr_gr, coln = 1)
}
