% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_split_data_basedon_annotation.r
\name{split_train_valid}
\alias{split_train_valid}
\title{Split Data into Training and Validation Sets}
\usage{
split_train_valid(df, ratio_valid = 0.2)
}
\arguments{
\item{df}{A data frame containing the data to be split, with columns
including `seqnames` and `txType`.}

\item{ratio_valid}{A numeric value specifying the proportion of the data
to be used as the validation set. Defaults to 0.2.}
}
\value{
A `GRangesList` object containing two `GRanges` objects: one for
  the training set and one for the validation set.
}
\description{
This function splits a given data frame into training and validation sets
based on a specified validation ratio. The split is stratified by
`seqnames` (chromosomes) and `txType` (transcript type).
}
\examples{
df <- data.frame(seqnames = rep("chr1", 100), txType = rep("gene", 100), start = 1:100, end = 101:200)
split_sets <- split_train_valid(df, ratio_valid = 0.2)
}
