% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.r
\name{strands_norm_subtraction}
\alias{strands_norm_subtraction}
\title{Normalized Strand Subtraction}
\usage{
strands_norm_subtraction(vec, len_vec)
}
\arguments{
\item{vec}{A numeric vector containing the combined forward and reverse 
strand signals.}

\item{len_vec}{An integer specifying the length of the forward strand 
signal within `vec`.}
}
\value{
A numeric vector containing the normalized subtraction result of the
forward and reverse strand signals.
}
\description{
This function performs a normalized subtraction between forward (plus) and 
reverse (minus) strand signals.
}
