% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_helper.r
\name{strands_norm_subtraction_all}
\alias{strands_norm_subtraction_all}
\title{Apply Normalized Strand Subtraction to All Windows}
\usage{
strands_norm_subtraction_all(windows, ext_dis, len_vec)
}
\arguments{
\item{windows}{A data frame where each row represents a window containing
combined forward and reverse strand signals.}

\item{ext_dis}{An integer specifying the extension distance,
used for adjusting column names.}

\item{len_vec}{An integer specifying the length of the forward strand signal
within each window.}
}
\value{
A data frame containing the normalized subtraction results 
for all windows.

Apply normalized forward and reverse subtraction to all windows
}
\description{
This function applies the normalized subtraction between forward (plus) and
reverse (minus) strand signals to all windows.
}
