% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_split_data_basedon_annotation.r
\name{subset_annotation_positive}
\alias{subset_annotation_positive}
\title{Subset Positive Annotations}
\usage{
subset_annotation_positive(pos_sltocr_grl)
}
\arguments{
\item{pos_sltocr_grl}{A `GRangesList` object containing positive selected 
open chromatin regions (OCRs) with a metadata column `txType` indicating 
transcript types.}
}
\value{
A `GRangesList` object with entries filtered to exclude those with 
  `txType` values 'exon', 'CDS', and 'threeUTR'.
}
\description{
This function subsets positive annotations from a `GRangesList` object by 
removing entries with specific transcript types (`txType`), such as 'exon', 
'CDS', and 'threeUTR'.
}
\examples{
# Example usage:
# pos_sltocr_grl is a GRangesList object with a 'txType' metadata column
filtered_grl <- subset_annotation_positive(pos_sltocr_grl)
}
