% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonoverlapping_ranges.r
\name{wrapup_filter_bed_to_reduce}
\alias{wrapup_filter_bed_to_reduce}
\title{Process a BED file and save the highest non-overlapping ranges}
\usage{
wrapup_filter_bed_to_reduce(input_bed, output_dir = NULL)
}
\arguments{
\item{input_bed}{Character. The path to the input BED file.}

\item{output_dir}{Character. The directory where the output files will be saved.}
}
\value{
A `GRanges` object containing the highest non-overlapping ranges.
}
\description{
This function loads a BED file, converts it to a `GRanges` object, and selects the highest
non-overlapping ranges using the `get_highest_non_overlap` function. The results are saved
as both RDS and BED files.
}
