% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.r
\name{wrapup_make_profiles}
\alias{wrapup_make_profiles}
\title{Wrap up profile for input of the model}
\usage{
wrapup_make_profiles(
  ctss_rse,
  regions_gr,
  output_dir,
  output_dir_name,
  output_subdir_name,
  ext_dis,
  addtn_to_filename = "",
  save_count_profiles = FALSE,
  file_type = "parquet"
)
}
\arguments{
\item{ctss_rse}{SummarizedExperiment object containing count data.}

\item{regions_gr}{GRanges or GRangesList object specifying genomic regions.}

\item{output_dir}{Directory where output profiles and metadata will be saved.}

\item{output_dir_name}{Output directory name within \code{dir_results}.}

\item{output_subdir_name}{Subdirectory name within \code{outdir_dir_name}
where output files will be stored.}

\item{ext_dis}{Numeric value specifying the extension distance.}

\item{addtn_to_filename}{Additional text to append to the output file names.}

\item{save_count_profiles}{Logical value indicating
whether to save count profiles.}

\item{file_type}{Character string indicating the file type for output 
("csv" or "parquet"). Default is "parquet".}
}
\description{
This function processes multiple columns of
data in \code{ctss_rse} and generates
count profiles and subtracted normalized profiles,
saving them along with metadata
as CSV or Parquet files in specified directories.
}
\details{
This function iterates through each column of \code{ctss_rse},
processes the data to generate count and subtracted normalized profiles,
and saves them as CSV or Parquet files.
It also creates corresponding metadata files based on
row names of the profiles.
}
\examples{
\dontrun{
# Example usage:
wrapup_make_profiles(ctss_rse, regions_gr, dir_results, outdir_dir_name,
                     outdir_subdir_name, ext_dis, file_type = "parquet")
}

}
