% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_split_data_basedon_annotation.r
\name{wrapup_split_train_valid}
\alias{wrapup_split_train_valid}
\title{Wrap Up Data Splitting into Training and Validation Sets}
\usage{
wrapup_split_train_valid(pos_grl, neg_grl)
}
\arguments{
\item{pos_grl}{A `GRangesList` object containing positive genomic ranges.}

\item{neg_grl}{A `GRangesList` object containing negative genomic ranges.}
}
\value{
A list containing four `GRangesList` objects: `pos_tr_grl` for
  positive training sets, `pos_vl_grl` for positive validation sets,
  `neg_tr_grl` for negative training sets, and `neg_vl_grl` for negative
  validation sets.
}
\description{
This function wraps up the process of splitting positive and negative
genomic ranges into training and validation sets. It ensures the number of
training and validation samples is balanced between positive and negative
sets.
}
\examples{
pos_grl <- GenomicRanges::GRangesList()
neg_grl <- GenomicRanges::GRangesList()
result <- wrapup_split_train_valid(pos_grl, neg_grl)
}
